requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.10.6
pydantic-settings==2.0.3
openai==1.3.0
anthropic==0.7.0
sqlparse==0.4.4
javalang==0.13.0
gitpython==3.1.40
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
asyncio-throttle==1.0.2
aiohttp==3.9.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1